# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DataForge API
  description: Immutable dataset versions, hashed provenance, conservative computed
    quality verdicts, deterministic drift, provenance closure, contract compatibility
    and downstream impact reports, with append-only audit.
  version: 1.0.7
servers:
- url: http://127.0.0.1:8010
paths:
...
  /v1/sources:
    post:
      tags:
      - sources
      summary: Create Source
      operationId: create_source_v1_sources_post
      parameters:
      - name: X-Actor
        in: header
        required: false
        schema:
          type: string
          default: api
          title: X-Actor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Source V1 Sources Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - sources
      summary: List Sources
      operationId: list_sources_v1_sources_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Sources V1 Sources Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/sources/{source_id}:
    get:
      tags:
      - sources
      summary: Get Source
      operationId: get_source_v1_sources__source_id__get
      parameters:
      - name: source_id
        in: path
        required: true
        schema:
          type: string
          title: Source Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Source V1 Sources  Source Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/datasets:
    post:
      tags:
      - datasets
      summary: Create Dataset
      operationId: create_dataset_v1_datasets_post
      parameters:
      - name: X-Actor
        in: header
        required: false
        schema:
          type: string
          default: api
          title: X-Actor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasetCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Dataset V1 Datasets Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - datasets
      summary: List Datasets
      operationId: list_datasets_v1_datasets_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Datasets V1 Datasets Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/datasets/{dataset_id}:
    get:
      tags:
      - datasets
      summary: Get Dataset
      operationId: get_dataset_v1_datasets__dataset_id__get
      parameters:
      - name: dataset_id
        in: path
        required: true
        schema:
          type: string
          title: Dataset Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Dataset V1 Datasets  Dataset Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/datasets/{dataset_id}/versions:
    post:
      tags:
      - versions
      summary: Create Version
      operationId: create_version_v1_datasets__dataset_id__versions_post
      parameters:
      - name: dataset_id
        in: path
        required: true
        schema:
          type: string
          title: Dataset Id
      - name: X-Actor
        in: header
        required: false
        schema:
          type: string
          default: api
          title: X-Actor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Version V1 Datasets  Dataset Id  Versions Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - versions
      summary: List Versions
      operationId: list_versions_v1_datasets__dataset_id__versions_get
      parameters:
      - name: dataset_id
        in: path
        required: true
        schema:
          type: string
          title: Dataset Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: include_records
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Records
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Versions V1 Datasets  Dataset Id  Versions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/versions/{version_id}:
    get:
      tags:
      - versions
      summary: Get Version
      operationId: get_version_v1_versions__version_id__get
      parameters:
      - name: version_id
        in: path
        required: true
        schema:
          type: string
          title: Version Id
      - name: include_records
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Include Records
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Version V1 Versions  Version Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/versions/{version_id}/provenance/verify:
    get:
      tags:
      - provenance
      summary: Verify Provenance
      operationId: verify_provenance_v1_versions__version_id__provenance_verify_get
      parameters:
      - name: version_id
        in: path
        required: true
        schema:
          type: string
          title: Version Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Verify Provenance V1 Versions  Version Id  Provenance
                  Verify Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/versions/{version_id}/quality:
    get:
      tags:
      - quality
      summary: Get Quality
      operationId: get_quality_v1_versions__version_id__quality_get
      parameters:
      - name: version_id
        in: path
        required: true
        schema:
          type: string
          title: Version Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Quality V1 Versions  Version Id  Quality Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/versions/{version_id}/quality-checks:
    post:
      tags:
      - quality
      summary: Recheck Quality
      operationId: recheck_quality_v1_versions__version_id__quality_checks_post
      parameters:
      - name: version_id
        in: path
        required: true
        schema:
          type: string
          title: Version Id
      - name: X-Actor
        in: header
        required: false
        schema:
          type: string
          default: api
          title: X-Actor
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Recheck Quality V1 Versions  Version Id  Quality Checks
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/versions/{baseline_version_id}/drift-reports:
    post:
      tags:
      - drift
      summary: Create Drift Report
      operationId: create_drift_report_v1_versions__baseline_version_id__drift_reports_post
      parameters:
      - name: baseline_version_id
        in: path
        required: true
        schema:
          type: string
          title: Baseline Version Id
      - name: X-Actor
        in: header
        required: false
        schema:
          type: string
          default: api
          title: X-Actor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriftComparisonCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Drift Report V1 Versions  Baseline Version
                  Id  Drift Reports Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/drift-reports/{report_id}:
    get:
      tags:
      - drift
      summary: Get Drift Report
      operationId: get_drift_report_v1_drift_reports__report_id__get
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: string
          title: Report Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Drift Report V1 Drift Reports  Report Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/datasets/{dataset_id}/drift-reports:
    get:
      tags:
      - drift
      summary: List Drift Reports
      operationId: list_drift_reports_v1_datasets__dataset_id__drift_reports_get
      parameters:
      - name: dataset_id
        in: path
        required: true
        schema:
          type: string
          title: Dataset Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Drift Reports V1 Datasets  Dataset Id  Drift
                  Reports Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/contracts/{contract_id}:
    get:
      tags:
      - contracts
      summary: Get Data Contract
      operationId: get_data_contract_v1_contracts__contract_id__get
      parameters:
      - name: contract_id
        in: path
        required: true
        schema:
          type: string
          title: Contract Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Data Contract V1 Contracts  Contract Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DatasetCreate:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 4000
          - type: 'null'
          title: Description
        schema:
          anyOf:
          - $ref: '#/components/schemas/DatasetSchema'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - name
      title: DatasetCreate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SourceCreate:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        kind:
          type: string
          maxLength: 100
          minLength: 1
          title: Kind
        uri:
          anyOf:
          - type: string
            maxLength: 2048
          - type: 'null'
          title: Uri
        description:
          anyOf:
          - type: string
            maxLength: 4000
          - type: 'null'
          title: Description
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
      additionalProperties: false
      type: object
      required:
      - name
      - kind
      title: SourceCreate
...